// src/models/player.model.ts
export interface Player {
  id: number;
  name: string;
  position: string;
  jerseyNumber: number;
  teamId: number;
  teamName: string;
  teamCity: string;
  teamAbbreviation: string;
  isStarter: boolean;
  gamesPlayed: number;
  minutesPerGame: number;
  points: number;
  rebounds: number;
  assists: number;
  steals: number;
  blocks: number;
  turnovers: number;
  fieldGoalPercentage: number;
  threePointPercentage: number;
  freeThrowPercentage: number;
  offensiveRebounds: number;
  defensiveRebounds: number;
  fieldGoalsMade: number;
  fieldGoalsAttempted: number;
  threePointersMade: number;
  threePointersAttempted: number;
  freeThrowsMade: number;
  freeThrowsAttempted: number;
  plusMinus: number;
  fantasyPoints: number;
  doubleDoubles: number;
  tripleDoubles: number;
  personalFouls: number;
  age: number;
  height: string;
  weight: number;
  efficiencyRating: number;
  trueShootingPercentage: number;
  effectiveFieldGoalPercentage: number;
  assistToTurnoverRatio: number;
  impactScore: number;
  usageRate: number;
  playerEfficiencyRating: number;
}